+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- computation of `delta_row_col` in `sidd_chip_to_html._proj_ecef_to_image`
+
 
 ## [0.2.0] - 2026-06-04